refactor: split out serializer module
